% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMonacoX.R
\name{getMonacoThemes}
\alias{getMonacoThemes}
\title{Monaco themes}
\usage{
getMonacoThemes()
}
\description{
Get the list of available themes of the Monaco editor. All
  themes are dark, excepted \code{"vs"}.
}
