% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monaco.R
\name{monaco-shiny}
\alias{monaco-shiny}
\alias{monacoOutput}
\alias{renderMonaco}
\title{Shiny bindings for Monaco editor}
\usage{
monacoOutput(outputId, width = "100\%", height = "400px")

renderMonaco(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{CSS measurements (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended}

\item{expr}{an expression that creates a Monaco editor with
\code{\link{monaco}}}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{logical, whether \code{expr} a quoted expression}
}
\description{
Output and render functions for using Monaco editors within Shiny
applications and interactive Rmd documents.
}
\examples{
library(monaco)
library(shiny)

ui <- fluidPage(
  monacoOutput("ed", height = "400px")
)

server <- function(input, output){

  output[["ed"]] <- renderMonaco({
    monaco(
      system.file("exampleFiles", "JavaScript.js", package = "monaco")
    )
  })

}

if(interactive()){
  shinyApp(ui, server)
}
}
