% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monaco.R
\name{monaco}
\alias{monaco}
\title{Monaco editor}
\usage{
monaco(
  contents,
  language = NULL,
  theme = NULL,
  tabSize = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{contents}{this can be the path to a file, \code{NULL} to open an
empty editor, or missing to open the file currently open in RStudio}

\item{language}{the language of the contents; if \code{NULL} and the
contents are read from a file, the mode is guessed from the extension of
the file; run \code{\link{getMonacoLanguages}} to get the list of
available languages}

\item{theme}{the theme of the editor; run \code{\link{getMonacoThemes}} to
get the list of available themes}

\item{tabSize}{number of spaces for the indentation (usually \code{2} or
\code{4}); if \code{NULL}, it is set to the one used in RStudio}

\item{width, height}{dimensions; the default values are nice for usage in
the RStudio viewer pane}

\item{elementId}{a HTML id for the container; this is useless for common
usage}
}
\description{
Open the Monaco editor.
}
\examples{
# in RStudio, `monaco()` opens the current file:
monaco()

# opens a new JavaScript file:
monaco(NULL, language = "javascript")

# opens an existing file:
monaco(system.file("exampleFiles", "JavaScript.js", package = "monaco"))

# try the SVG viewer; you can zoom and pan the image:
monaco(system.file("exampleFiles", "react.svg", package = "monaco"))
}
